Added Key callback and Error callback functions
Added static void key_callback and error_callback functions.  The key
callback demonstrates and tests glfw key callback events.  In this case,
it detects when the key 'Escape' is pressed and shuts down the program.
Error callback sends errors to stderr and is able to give errors before
glfw intitialization.