Reorganized dir structure and edited vcxproj
Directory structure has been reorganized and VS 2010 is set to mirror this directory structure, as well as place object files in a mirrored directory structure.  Dir structure will change more over the life of the project as engine becomes more seperated from game and other projects are added to the solution.